fix: replace library cmath Exercise 5

+++ Lab1/src/guia1.h
@@ -3,7 +3,7 @@
 
 class GuiaLab {
 public:
-  int get_remainder(int a, int b);
+  float get_remainder(int a, int b);
 
   const char *is_even_or_odd(int number);
